Handle JWT token expiration
