+++ src/app/App.tsx
@@ -1,23 +1,41 @@
 import React from "react";
 
-import { DemoBanner } from "@components/atoms";
+import { useAuth } from "@saleor/sdk";
+import { DemoBanner, Loader } from "@components/atoms";
 import { demoMode } from "@temp/constants";
-import { Footer, MainMenu, MetaConsumer, OverlayManager } from "../components";
+import {
+  Footer,
+  MainMenu,
+  MetaConsumer,
+  OverlayManager,
+  OverlayProvider,
+} from "../components";
+import ShopProvider from "../components/ShopProvider";
 import "../globalStyles/scss/index.scss";
 import { Routes } from "./routes";
+import Notifications from "./Notifications";
 
 const App: React.FC = () => {
+  const { tokenRefreshing } = useAuth();
+
+  if (tokenRefreshing) {
+    return <Loader />;
+  }
+
   return (
-    <>
-      <MetaConsumer />
-      {demoMode && <DemoBanner />}
-      <header>
-        <MainMenu />
-      </header>
-      <Routes />
-      <Footer />
-      <OverlayManager />
-    </>
+    <ShopProvider>
+      <OverlayProvider>
+        <MetaConsumer />
+        {demoMode && <DemoBanner />}
+        <header>
+          <MainMenu />
+        </header>
+        <Routes />
+        <Footer />
+        <OverlayManager />
+        <Notifications />
+      </OverlayProvider>
+    </ShopProvider>
   );
 };
